+++ packages/dictionary/test/dictionary-test.js
@@ -1,23 +1,30 @@
-var tape = require('tape')
+var test = require('tape')
 var dict = require('..')
 
 var DATA = {
-  'maj7': ['1 3 5 7', ['Maj7']],
-  'm7': ['1 b3 5 7']
+  'maj7': ['1P 3M 5P 7M', ['Maj7']],
+  'm7': ['1P 3m 5P 7m']
 }
 
-tape('dictionary: get', function (test) {
-  var get = dict.get(null, DATA)
-  test.deepEqual(get('maj7'), [ '1', '3', '5', '7' ])
-  test.deepEqual(get('Maj7'), [ '1', '3', '5', '7' ])
-  test.deepEqual(get('m7'), [ '1', 'b3', '5', '7' ])
-  test.equal(get('blah'), undefined)
-  test.end()
+test('dictionary: detector', function (t) {
+  t.deepEqual(dict.detector(null, DATA)('E4 C4 B2 G'), [ ['maj7', 'G'] ])
+  t.deepEqual(dict.detector('', DATA)('E4 C4 B2 G'), [ 'Gmaj7' ])
+  t.deepEqual(dict.detector(' ', DATA)('E4 C4 B2 G'), [ 'G maj7' ])
+  t.end()
 })
 
-tape('dictionary: keys', function (test) {
+test('dictionary: get', function (t) {
+  var get = dict.get(null, DATA)
+  t.deepEqual(get('maj7'), [ '1P', '3M', '5P', '7M' ])
+  t.deepEqual(get('Maj7'), [ '1P', '3M', '5P', '7M' ])
+  t.deepEqual(get('m7'), [ '1P', '3m', '5P', '7m' ])
+  t.equal(get('blah'), undefined)
+  t.end()
+})
+
+test('dictionary: keys', function (t) {
   var keys = dict.keys(DATA)
-  test.deepEqual(keys(), [ 'maj7', 'm7' ])
-  test.deepEqual(keys(true), [ 'maj7', 'm7', 'Maj7' ])
-  test.end()
+  t.deepEqual(keys(), [ 'maj7', 'm7' ])
+  t.deepEqual(keys(true), [ 'maj7', 'm7', 'Maj7' ])
+  t.end()
 })